change uint64 to int

+++ projects/activitytracker1/pkg/server/activity.go
@@ -10,21 +10,21 @@ var ErrIDNotFound = errors.New("ID not found")
 type Activity struct {
 	Time        time.Time `json:"time"`
 	Description string    `json:"description"`
-	ID          uint64    `json:"id"`
+	ID          int       `json:"id"`
 }
 
 type Activities struct {
 	Activities []Activity
 }
 
-func (a *Activities) Insert(activity Activity) uint64 {
-	activity.ID = uint64(len(a.Activities))
+func (a *Activities) Insert(activity Activity) int {
+	activity.ID = (len(a.Activities))
 	a.Activities = append(a.Activities, activity)
 	return activity.ID
 }
 
-func (a *Activities) Retrieve(id uint64) (Activity, error) {
-	if id >= uint64(len(a.Activities)) {
+func (a *Activities) Retrieve(id int) (Activity, error) {
+	if id >= int(len(a.Activities)) {
 		return Activity{}, ErrIDNotFound
 	}
 	return a.Activities[id], nil
